[TRL-171] test: 테스트 코드 패키지 구조 변경
