--- src/test/java/com/cosain/trilo/trip/command/adapter/in/api/scheduleplace/SchedulePlaceDeleteControllerTest.java
@@ -1,56 +1,0 @@
-package com.cosain.trilo.trip.command.adapter.in.api.scheduleplace;
-
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.security.test.context.support.WithAnonymousUser;
-import org.springframework.security.test.context.support.WithMockUser;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-import org.springframework.web.context.WebApplicationContext;
-
-import static org.springframework.security.test.web.servlet.setup.SecurityMockMvcConfigurers.springSecurity;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
-import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-@SpringBootTest
-@AutoConfigureMockMvc
-@DisplayName("일정장소 삭제 API 테스트")
-class SchedulePlaceDeleteControllerTest {
-
-    @Autowired
-    private MockMvc mockMvc;
-
-    @BeforeEach
-    void setUp(WebApplicationContext context) {
-        mockMvc = MockMvcBuilders.webAppContextSetup(context)
-                .apply(springSecurity())
-                .build();
-    }
-
-    @Test
-    @DisplayName("인증된 사용자 요청 -> 미구현 500")
-    @WithMockUser
-    public void deleteSchedulePlace_with_authorizedUser() throws Exception {
-        mockMvc.perform(delete("/api/schedule-places/1"))
-                .andDo(print())
-                .andExpect(content().string("[일정장소 삭제] This operation is not implemented."))
-                .andExpect(status().isInternalServerError());
-    }
-
-    @Test
-    @DisplayName("미인증 사용자 요청 -> 인증 실패 401")
-    @WithAnonymousUser
-    public void deleteSchedulePlace_with_unauthorizedUser() throws Exception {
-        mockMvc.perform(delete("/api/schedule-places/1"))
-                .andDo(print())
-                .andExpect(content().string(""))
-                .andExpect(status().isUnauthorized());
-    }
-
-}
